dev-requirements.txt
beautifulsoup4
factory-boy>=2
mock
pytest
pytest-ckan
pytest-cov
